Code Formatting Extracted LOADER and VT into their own file moved MVDescriptor and PackAction into Pack module moved confirm_input and read_line_to_string to own module
